Refactor plugin_interface functions to static methods
