+++ libraries/classes/Display/Export.php
@@ -10,6 +10,7 @@ namespace PhpMyAdmin\Display;
 use PhpMyAdmin\Core;
 use PhpMyAdmin\Encoding;
 use PhpMyAdmin\Message;
+use PhpMyAdmin\Plugins;
 use PhpMyAdmin\Plugins\ExportPlugin;
 use PhpMyAdmin\Relation;
 use PhpMyAdmin\Response;
@@ -371,7 +372,7 @@ class Export
     {
         $html  = '<div class="exportoptions" id="format">';
         $html .= '<h3>' . __('Format:') . '</h3>';
-        $html .= PMA_pluginGetChoice('Export', 'what', $export_list, 'format');
+        $html .= Plugins::getChoice('Export', 'what', $export_list, 'format');
         $html .= '</div>';
         return $html;
     }
@@ -393,7 +394,7 @@ class Export
             . 'and ignore the options for other formats.'
         );
         $html .= '</p>';
-        $html .= PMA_pluginGetOptions('Export', $export_list);
+        $html .= Plugins::getOptions('Export', $export_list);
         $html .= '</div>';
 
         if (Encoding::canConvertKanji()) {
@@ -1002,11 +1003,9 @@ class Export
             $GLOBALS['single_table'] = $_REQUEST['single_table'];
         }
 
-        include_once './libraries/plugin_interface.lib.php';
-
         /* Scan for plugins */
         /* @var $export_list ExportPlugin[] */
-        $export_list = PMA_getPlugins(
+        $export_list = Plugins::getPlugins(
             "export",
             'libraries/classes/Plugins/Export/',
             array(
